% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{EvoResistanceNHEJ-package}
\alias{EvoResistanceNHEJ}
\alias{EvoResistanceNHEJ-package}
\title{EvoResistanceNHEJ}
\description{
Functions included in this package allow for the modeling of driver allele spread as proposed by Unckless and Messer 2015. This variation explicitly accounts for errors generated during NHEJ.
}
\author{
\strong{Maintainer}: Jared Bennett \email{temp@gmail.com}

Authors:
\itemize{
  \item Sean Wu \email{slwu89@berkeley.edu}
}

}
