% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{fixation_probability}
\alias{fixation_probability}
\title{fixation_probability}
\usage{
fixation_probability(P, T)
}
\arguments{
\item{P:}{dunno what this is}

\item{T:}{dunno what this is}
}
\description{
fixation_probability
}
