% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_delta}
\alias{p_delta}
\title{Probability of resistance through NHEJ.}
\usage{
p_delta(P, T, F)
}
\arguments{
\item{P}{Parameter object. Create this using the "parameters()" constructor.}

\item{T}{Trajectory object. Create this using the "trajectory()" constructor.}

\item{F}{Fixation_probability object. Create this using the "fixation_probability()" constructor.}
}
\description{
p_delta calculates the probability of resistance from NHEJ during driver homing. This is the probability of any mutation being resistant in t<tfix.
}
