% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{p_sgv}
\alias{p_sgv}
\title{Probability of resistance through standing genetic variation.}
\usage{
p_sgv(P, F)
}
\arguments{
\item{P}{Parameter object. Create this using the "parameters()" constructor.}

\item{F}{Fixation_probability object. Create this using the "fixation_probability()" constructor.}
}
\description{
p_sgv calculates the probability of resistance existing as standing genetic variation. This is the probability of resistance becoming fixed during t<tfix.
}
