% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppR6.R
\name{fixation_probability}
\alias{fixation_probability}
\title{Fixation_probability object for EvoResistanceNHEJ simulations.}
\usage{
fixation_probability(P, T)
}
\arguments{
\item{P:}{Parameter object. Create this using the "parameters()" constructor.}

\item{T:}{Trajectory object. Create this using the "trajectory()" constructor.}
}
\description{
This class calculates the fixation probability of the driver allele.
}
